requirements.txt
fastapi
paramiko
pydantic
uvicorn
simplejson
mysql-connector
mysql-connector-python
pyjwt
redis